Nettoyage du code, astyle...

+++ src/FenetreJeu.java
@@ -3,16 +3,12 @@ import java.awt.*;
 import java.awt.event.*;
 import javax.swing.event.*;
 
-public class FenetreJeu extends JPanel implements KeyListener 
+public class FenetreJeu extends JPanel implements KeyListener
 {
-	final int duree = 10 ;
+    final int duree = 30;
     JeuCalcul jeu;
     FenetreModele fm;
-    JFrame fenetre;
-
-    JMenuItem mReset;
-    JMenuItem mJouer;
-
+    //JFrame fenetre;
 
     JPanel pBouton;
     JButton bJouer;
@@ -25,8 +21,7 @@ public class FenetreJeu extends JPanel implements KeyListener
     int temps=duree;
     JLabel textTemps;
 
-    JPanel pJeu = new JPanel(new BorderLayout());
-    JPanel pExpression;
+    JPanel pJeu = new JPanel();
     JLabel labelExpression;
     JLabel labelResultat;
     ImageIcon iResultatTrouve;
@@ -42,90 +37,53 @@ public class FenetreJeu extends JPanel implements KeyListener
     {
         this.fm = f;
         this.fm.setTitre("Calcul Mental - Jeu");
-        fenetre = fm.getJFrame();
-
-        /*Creation du menu
-          *
-          */
-        /* menu Jeu */
-       /* mJeu = new JMenu("Jeu");
-        mJeu.setFont(fCambMini);
-        mQuitter = new JMenuItem("Quitter");
-        mQuitter.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_W,
-                                KeyEvent.CTRL_MASK));
-        mQuitter.addActionListener(new ActionListener() {
-            public void actionPerformed(ActionEvent arg0) {
-                System.exit(0);
-            }
-        });
-        mJouer = new JMenuItem("Jouer");
-        mJouer.setFont(fCambMini);
-        mJouer.addActionListener(new ActionJouer());
-
-        mReset = new JMenuItem("Nouvelle Partie");
-        mReset.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_N,
-                              KeyEvent.CTRL_MASK));
-        mReset.addActionListener(new ActionReset());
-        mJeu.add(mJouer);
-        mJeu.add(mReset);
-        mJeu.add(mQuitter);
-        menuBar.add(mJeu);*/
 
         /*
          * Initialisation des panel
          */
         pBouton = new JPanel();
-        pExpression = new JPanel();
         pInfo = new JPanel();
         pChrono = new JPanel();
 
         /* Creation des bouton */
         bJouer = new JButton("Commencer");
         bJouer.setFont(FenetreModele.fCamb);
-        pBouton.add(bJouer);
         bJouer.addActionListener(new ActionJouer());
+        pBouton.add(bJouer);
 
-        bReset = new JButton("Arrêter");
+        bReset = new JButton("Retour");
         bReset.setFont(FenetreModele.fCamb);
-        bReset.setEnabled(false);
-
-        pBouton.add(bReset);
+        bReset.setEnabled(true);
         bReset.addActionListener(new ActionReset());
+        pBouton.add(bReset);
 
         bQuitter = new JButton("Quitter");
         bQuitter.setFont(FenetreModele.fCamb);
-        pBouton.add(bQuitter);
         bQuitter.addActionListener(new ActionQuitter());
-
-        fenetre.getContentPane().add(pBouton, BorderLayout.SOUTH);
+        pBouton.add(bQuitter);
 
         jeu = new JeuCalcul(n);
-
         /* Panneau contenant l'expression et la saisie */
         //labelExpression = new JLabel (jeu.retournerExpression().toString());
-        labelExpression = new JLabel ("");
+        labelExpression = new JLabel ();
         labelExpression.setPreferredSize( new Dimension( 80, 40 ) );
-        pExpression.add(labelExpression);
+        pJeu.add(labelExpression);
 
         reponse = jeu.retournerExpression().solution();
         labelReponse = new JTextField();
         labelReponse.setPreferredSize( new Dimension( 40, 24 ) );
         labelReponse.addKeyListener(this);
         labelReponse.setEnabled(false);
-        pExpression.add(labelReponse);
+        pJeu.add(labelReponse);
 
 
-        labelResultat = new JLabel("");
+        labelResultat = new JLabel();
         labelResultat.setPreferredSize(new Dimension(64,48));
         labelResultat.setHorizontalAlignment(SwingConstants.RIGHT);
-        pExpression.add(labelResultat);
+        pJeu.add(labelResultat);
 
         iResultatTrouve = new ImageIcon("img/right.png");
         iResultatFaux = new ImageIcon("img/wrong.png");
-
-        pJeu.add(pExpression,BorderLayout.CENTER);
-        //fenetre.getContentPane().add(pExpression, BorderLayout.CENTER);
-        fenetre.getContentPane().add(pJeu, BorderLayout.CENTER);
 
         /*Panneau du score*/
         JLabel textScore = new JLabel("Score : ");
@@ -136,26 +94,28 @@ public class FenetreJeu extends JPanel implements KeyListener
         labelScore.setPreferredSize(new Dimension(42,24));
         pInfo.add(labelScore);
 
-        fenetre.getContentPane().add(pInfo, BorderLayout.EAST);
-
-        /* Compte a rebour*/
+        /* Compte a rebours*/
         textTemps = new JLabel("Temps Restant : ");
         textTemps.setFont(FenetreModele.fCamb);
         labelTemps = new JLabel(Integer.toString(temps));
         labelTemps.setFont(FenetreModele.fCamb);
         pChrono.add(textTemps);
         pChrono.add(labelTemps);
-        fenetre.getContentPane().add(pChrono, BorderLayout.NORTH);
 
-
-        fenetre.pack();
-        fenetre.setVisible(true);
+        this.fm.getContentPane().setLayout(new BorderLayout());
+        this.add(pChrono, BorderLayout.NORTH);
+        this.add("Center",Box.createRigidArea(new Dimension(40,0)));
+        this.add(pInfo, BorderLayout.EAST);
+        this.add(pJeu, BorderLayout.CENTER);
+        this.add(pBouton, BorderLayout.SOUTH);
     }
 
-    public void keyTyped(KeyEvent e) {
+    public void keyTyped(KeyEvent e)
+    {
     }
 
-    public void keyPressed(KeyEvent e) {
+    public void keyPressed(KeyEvent e)
+    {
 
         if (e.getKeyCode() == KeyEvent.VK_ENTER)
         {
@@ -176,10 +136,10 @@ public class FenetreJeu extends JPanel implements KeyListener
                     labelResultat.setIcon(iResultatFaux);
                     labelSolution.setForeground(Color.red);
                     labelSolution.setText( "La solution de " + jeu.retournerExpression().toString2() + " est " + reponse  );
-					pJeu.add(labelSolution, BorderLayout.SOUTH);
-					labelSolution.setPreferredSize(new Dimension(170,30));
+                    pJeu.add(labelSolution, BorderLayout.SOUTH);
+                    labelSolution.setPreferredSize(new Dimension(170,30));
                 }
-                
+
                 jeu.supprimerExpression();
                 labelExpression.setText(jeu.retournerExpression().toString());
                 reponse = jeu.retournerExpression().solution();
@@ -187,7 +147,7 @@ public class FenetreJeu extends JPanel implements KeyListener
             }
             catch(NumberFormatException exc)
             {
-                JOptionPane.showMessageDialog(fenetre,"Un nombre est attendu comme réponse","Erreur", JOptionPane.ERROR_MESSAGE);
+                JOptionPane.showMessageDialog(this,"Un nombre est attendu comme réponse","Erreur", JOptionPane.ERROR_MESSAGE);
             }
             finally
             {
@@ -200,13 +160,20 @@ public class FenetreJeu extends JPanel implements KeyListener
 
     }
 
-    public void keyReleased(KeyEvent e) {
+    public void keyReleased(KeyEvent e)
+    {
     }
 
     class ActionQuitter implements ActionListener
     {
         public void actionPerformed(ActionEvent ec)
         {
+            if ( horloge != null )
+            {
+                horloge.cancel();
+                horloge.purge();
+                horloge = null;
+            }
             System.exit(0);
         }
     }
@@ -215,28 +182,35 @@ public class FenetreJeu extends JPanel implements KeyListener
     {
         public void actionPerformed(ActionEvent ec)
         {
-            jeu.razScore();
-            labelScore.setText(jeu.getScore()+"");
-
-            labelSolution.setText("");
-            jeu.supprimerExpression();
-            labelResultat.setIcon(null);
-            //labelExpression.setText(jeu.retournerExpression().toString());
-            labelExpression.setText("");
-            reponse = jeu.retournerExpression().solution();
-            labelReponse.setText("");
-            if(horloge != null)
+            if(bReset.getText().equals("Arrêter"))
             {
-                horloge.cancel();
-                horloge.purge();
-                horloge= null;
-            }
-            if( !bJouer.isEnabled())
-                bJouer.setEnabled(true);
+                jeu.razScore();
+                labelScore.setText(jeu.getScore()+"");
 
-			bReset.setEnabled(false);
-            temps = duree;
-            labelTemps.setText(temps+"");
+                labelSolution.setText("");
+                jeu.supprimerExpression();
+                labelResultat.setIcon(null);
+                //labelExpression.setText(jeu.retournerExpression().toString());
+                labelExpression.setText("");
+                reponse = jeu.retournerExpression().solution();
+                labelReponse.setText("");
+                if(horloge != null)
+                {
+                    horloge.cancel();
+                    horloge.purge();
+                    horloge= null;
+                }
+                if( !bJouer.isEnabled())
+                    bJouer.setEnabled(true);
+
+                bReset.setText("Retour");
+                temps = duree;
+                labelTemps.setText(temps+"");
+            }
+            else
+            {
+                fm.menu();
+            }
         }
     }
 
@@ -251,16 +225,34 @@ public class FenetreJeu extends JPanel implements KeyListener
                 horloge = new java.util.Timer();
                 horloge.schedule( new Decompte(),1000,1000);
                 bJouer.setEnabled(false);
+                bReset.setText("Arrêter");
                 bReset.setEnabled(true);
             }
         }
     }
 
-
-    class Decompte extends java.util.TimerTask {
-        public void run() {
+    class Decompte extends java.util.TimerTask
+    {
+        public void run()
+        {
             if(temps !=0)
-                temps--;
+                if (temps > duree*3/4)
+                {
+                    labelTemps.setForeground(FenetreModele.VERT);
+                }
+                else if (temps > duree*1/2)
+                {
+                    labelTemps.setForeground(FenetreModele.JAUNE);
+                }
+                else if (temps > duree*1/4)
+                {
+                    labelTemps.setForeground(FenetreModele.ORANGE);
+                }
+                else
+                {
+                    labelTemps.setForeground(FenetreModele.ROUGE);
+                }
+            temps--;
             if (temps <= 0)
             {
                 //labelReponse.disable();
@@ -275,5 +267,5 @@ public class FenetreJeu extends JPanel implements KeyListener
             labelTemps.setText(temps+"");
         }
     }
-    
+
 }
